Move MakeTxParams to common types
